drone-search-engine/backend/requirements.txt
# Backend API
fastapi>=0.115.6,<1.0.0
uvicorn[standard]>=0.30.6,<1.0.0
python-dotenv>=1.0.1,<2.0.0

# AI Models
easyocr==1.7.1
transformers==4.44.2
accelerate==0.33.0
ultralytics==8.2.0
opencv-python==4.10.0.84
open-clip-torch==2.26.1
torch>=2.4.0
torchvision>=0.19.0

# Vector DB
qdrant-client==1.11.0

# Data / Math
numpy>=1.26.0,<3.0.0
pillow>=10.4.0,<12.0.0
scikit-learn==1.5.1

# Report Generation
groq==0.11.0
reportlab==4.2.2

# Utilities
python-multipart>=0.0.18,<1.0.0
cloudinary>=1.41.0,<2.0.0
